+++ packages/apalis-core/src/lib.rs
@@ -62,6 +62,9 @@ pub mod task;
 /// Codec for handling data
 pub mod codec;
 
+/// Allows stepped tasks
+pub mod step;
+
 /// Sleep utilities
 #[cfg(feature = "sleep")]
 pub async fn sleep(duration: std::time::Duration) {
@@ -211,9 +214,9 @@ pub mod test_utils {
 
     impl<B, Req, Res, Ctx> TestWrapper<B, Request<Req, Ctx>, Res>
     where
-        B: Backend<Request<Req, Ctx>, Res> + Send + Sync + 'static + Clone,
-        Req: Send + 'static + Sync,
-        Ctx: Send + Sync,
+        B: Backend<Request<Req, Ctx>> + Send + Sync + 'static + Clone,
+        Req: Send + 'static,
+        Ctx: Send,
         B::Stream: Send + 'static,
         B::Stream: Stream<Item = Result<Option<Request<Req, Ctx>>, crate::error::Error>> + Unpin,
         Res: Debug,
@@ -227,7 +230,7 @@ pub mod test_utils {
         Svc::Response: Send + Sync + 'static,
         Svc::Error: Send + Sync + std::error::Error,
         Ctx: Send + Sync + 'static,
-        B: Backend<Request<Req, Ctx>, Res> + 'static,
+        B: Backend<Request<Req, Ctx>> + 'static,
         B::Layer: Layer<Svc>,
         <B::Layer as Layer<Svc>>::Service: Service<Request<Req, Ctx>, Response = Res> + Send + Sync,
         B::Stream: Unpin + Send,
@@ -318,7 +321,7 @@ pub mod test_utils {
 
     impl<B, Req, Res, Ctx> Deref for TestWrapper<B, Request<Req, Ctx>, Res>
     where
-        B: Backend<Request<Req, Ctx>, Res>,
+        B: Backend<Request<Req, Ctx>>,
     {
         type Target = B;
 
@@ -329,7 +332,7 @@ pub mod test_utils {
 
     impl<B, Req, Ctx, Res> DerefMut for TestWrapper<B, Request<Req, Ctx>, Res>
     where
-        B: Backend<Request<Req, Ctx>, Res>,
+        B: Backend<Request<Req, Ctx>>,
     {
         fn deref_mut(&mut self) -> &mut Self::Target {
             &mut self.backend
@@ -598,6 +601,7 @@ pub mod test_utils {
                 let (mut t, poller) = TestWrapper::new_with_service(t.backend, service);
                 tokio::spawn(poller);
 
+                apalis_core::sleep(Duration::from_secs(1)).await;
                 // This is testing resuming the same worker
                 // This ensures that the worker resumed any jobs lost during an interruption
                 let res = t