+++ packages/apalis-cron/src/pipe.rs
@@ -4,7 +4,6 @@ use apalis_core::request::BoxStream;
 use apalis_core::{poller::Poller, request::Request, worker::Context, worker::Worker};
 use futures::StreamExt;
 use std::{error, fmt};
-use tower::Service;
 
 /// A generic Pipe that wraps an inner type along with a `RequestStream`.
 pub struct CronPipe<Inner> {
@@ -21,20 +20,19 @@ impl<Inner: fmt::Debug> fmt::Debug for CronPipe<Inner> {
     }
 }
 
-impl<T, Res, Ctx, Inner> Backend<Request<T, Ctx>, Res> for CronPipe<Inner>
+impl<T, Ctx, Inner> Backend<Request<T, Ctx>> for CronPipe<Inner>
 where
-    Inner: Backend<Request<T, Ctx>, Res>,
+    Inner: Backend<Request<T, Ctx>>,
 {
     type Stream = Inner::Stream;
 
     type Layer = Inner::Layer;
 
-    fn poll<Svc: Service<Request<T, Ctx>, Response = Res>>(
-        mut self,
-        worker: &Worker<Context>,
-    ) -> Poller<Self::Stream, Self::Layer> {
+    type Codec = Inner::Codec;
+
+    fn poll(mut self, worker: &Worker<Context>) -> Poller<Self::Stream, Self::Layer> {
         let pipe_heartbeat = async move { while (self.stream.next().await).is_some() {} };
-        let inner = self.inner.poll::<Svc>(worker);
+        let inner = self.inner.poll(worker);
         let heartbeat = inner.heartbeat;
 
         Poller::new_with_layer(